+++ oct_fmt.c
@@ -10,7 +10,7 @@
  */
 char *oct_fmt(va_list args, FMT *fmt)
 {
-	long n = va_arg(args, long);
+	long n = va_arg(args, int);
 	int i = 1;
 	char *num = malloc(i + 1);
